191.Research-Paper-extraction-Project-main/requirements.txt
transformers 
biopython
TensorFlow
torch
flask
streamlit
tf-keras